Make a thing's slug required.
(Get your head out of the gutter!)

For various reasons, it's desirable to make slug required. We need it
for saving data, most importantly, so requiring it only makes sense. In
Adventure, we still need to differentiate loading mission data into
memory versus restoring gamestate within an adventure.

Some tests and other minor changes were needed to support this.

If you don't know the slug for a Thing, call it new or test, and change
it at your earliest convenience.