Committing all WIP, hit a crossroads.
I am thinking that my ideas for dealing with changes in in-game things
is getting too complex. Rather than have a complex series of condition
checks, I am going to add the concept of states to things.

Every thing has a default state. A thing can have multiple states, and
can change between them when an event in the game makes it necessary to
do so. Every thing with carry a hashref of its state data, and will
refer to it (rather than rereading the YAML) when state needs to change.

This will break a lot, but I think the payoff will be worth it.